migrate storage layer to database

+++ src/main/java/ru/yandex/practicum/filmorate/storage/film/FilmStorage.java
@@ -1,6 +1,6 @@
 package ru.yandex.practicum.filmorate.storage.film;
 
-import ru.yandex.practicum.filmorate.model.Film;
+import ru.yandex.practicum.filmorate.model.film.Film;
 
 import java.util.Collection;
 import java.util.Optional;
@@ -11,10 +11,10 @@ public interface FilmStorage {
 
     Film addFilm(Film film);
 
-    Optional<Film> getFilmById(int id);
+    Optional<Film> getFilmById(long id);
 
     Film updateFilm(Film film);
 
-    boolean existsFilmById(Integer id);
+    boolean existsFilmById(long id);
 
 }
