+++ src/main/java/ru/yandex/practicum/filmorate/storage/film/InMemoryFilmStorage.java
@@ -1,7 +1,7 @@
 package ru.yandex.practicum.filmorate.storage.film;
 
 import org.springframework.stereotype.Component;
-import ru.yandex.practicum.filmorate.model.Film;
+import ru.yandex.practicum.filmorate.model.film.Film;
 
 
 import java.util.Collection;
@@ -11,8 +11,8 @@ import java.util.Optional;
 
 @Component
 public class InMemoryFilmStorage implements FilmStorage {
-    private final Map<Integer, Film> films = new HashMap<>();
-    private int nextId = 0;
+    private final Map<Long, Film> films = new HashMap<>();
+    private long nextId = 0;
 
     @Override
     public Collection<Film> getFilms() {
@@ -20,7 +20,7 @@ public class InMemoryFilmStorage implements FilmStorage {
     }
 
     @Override
-    public Optional<Film> getFilmById(int id) {
+    public Optional<Film> getFilmById(long id) {
         return Optional.ofNullable(films.get(id));
     }
 
@@ -38,11 +38,11 @@ public class InMemoryFilmStorage implements FilmStorage {
     }
 
     @Override
-    public boolean existsFilmById(Integer id) {
+    public boolean existsFilmById(long id) {
         return films.containsKey(id);
     }
 
-    private int getNextId() {
+    private long getNextId() {
         return nextId += 1;
     }
 }
