+++ src/main/java/ru/yandex/practicum/filmorate/storage/user/UserStorage.java
@@ -1,6 +1,6 @@
 package ru.yandex.practicum.filmorate.storage.user;
 
-import ru.yandex.practicum.filmorate.model.User;
+import ru.yandex.practicum.filmorate.model.user.User;
 
 import java.util.Collection;
 import java.util.Optional;
@@ -9,11 +9,11 @@ public interface UserStorage {
 
     Collection<User> getUsers();
 
-    Optional<User> getUserById(Integer userId);
+    Optional<User> getUserById(long userId);
 
     User addUser(User user);
 
     User updateUser(User user);
 
-    boolean existsUserById(Integer id);
+    boolean existsUserById(long id);
 }
